Fixed bugs
Hidden bug found. Occurred when patches first loaded do not contain any pixels of a class so that class wheel cannot be yielded from.

Solution: Add back in a init for wheels with a random row of data.

Note:  As each worker copies the same initialised DataLoader, this means that the initial wheels of all workers are the same